Split USERTYPE out to its own file
